서비스 제한 정책 로직 구현
- IDP 관련 로직 Deprecated 처리
-  tenantId 삭제

+++ storeplatform-sac/src/main/java/com/skplanet/storeplatform/sac/member/user/service/LimitTargetService.java
@@ -52,7 +52,7 @@ public interface LimitTargetService {
 	 * @param limitTarget 서비스 제한 정책
 	 * @return 저장 건수
 	 */
-	Long saveLimitPolicy(LimitTarget limitTarget);
+	Integer saveLimitPolicy(LimitTarget limitTarget);
 
 	/**
 	 * <pre>
